transit menu items to properties

+++ Mac/Sources/ColorPalette.m
@@ -153,7 +153,7 @@ static int color_remap [] =
     // 24 * 4 = 96
     BOOL remap = [dict count] == 24 * 4;
     
-    for (int i = 0; i < [self numberOfColors]; i++)
+    for (unsigned i = 0; i < [self numberOfColors]; i++)
     {
         int x = remap ? color_remap [i] : i;
 
@@ -182,21 +182,21 @@ static int color_remap [] =
 - (void)loadFromXChatFile:(int)file {
     struct stat filestat;
     fstat(file, &filestat);
-    char *cfg = malloc(filestat.st_size + 1);
+    char *cfg = malloc((size_t)filestat.st_size + 1);
     if (cfg != NULL) {
         cfg[0]  = '\0';
-        ssize_t cfglen = read(file, cfg, filestat.st_size);
+        ssize_t cfglen = read(file, cfg, (size_t)filestat.st_size);
         if (cfglen >= 0)
             cfg[cfglen] = '\0';
         
         int red, green, blue;
-        for (int i = 0; i < 32; i++) {
+        for (unsigned i = 0; i < 32; i++) {
             const char *name = [[NSString stringWithFormat:@"color_%d", i] UTF8String];
             cfg_get_color(cfg, (char *)name, &red, &green, &blue);
             [colors[i] release];
             colors[i] = [[NSColor colorWithDeviceRed:(CGFloat)red/0xffff green:(CGFloat)green/0xffff blue:(CGFloat)blue/0xffff alpha:1.0f] retain];
         }
-        for (int i = 256, j = 32; j < [self numberOfColors]; i++, j++) {
+        for (unsigned i = 256, j = 32; j < [self numberOfColors]; i++, j++) {
             const char *name = [[NSString stringWithFormat:@"color_%d", i] UTF8String];
             cfg_get_color(cfg, (char *)name, &red, &green, &blue);
             [colors[j] release];
@@ -235,7 +235,7 @@ static int color_remap [] =
 		}
         
 		/* our special colors are mapped at 256+ */
-		for (int i = 256, j = 32; j < self.numberOfColors; i++, j++)
+		for (unsigned i = 256, j = 32; j < self.numberOfColors; i++, j++)
 		{
 			const char *name = [[NSString stringWithFormat:@"color_%d", i] UTF8String];
 			cfg_put_color (file, colors[j].redComponent * 0xffff, colors[j].greenComponent * 0xffff, colors[j].blueComponent * 0xffff, (char *)name);
